+++ flake-module.nix
@@ -99,16 +99,55 @@ in
                     description = ''Package overrides given new source path'';
                     default = { };
                   };
-                  overrides = mkOption {
-                    type = functionTo (functionTo (types.lazyAttrsOf raw));
-                    description = ''
-                      Overrides for the Cabal project
+                  overrides =
+                    let
+                      # WARNING: While the order is deterministic, it is not
+                      # determined by the user. Thus overlays may be applied in
+                      # an unexpected order.
+                      # We need: https://github.com/NixOS/nixpkgs/issues/215486
+                      haskellOverlayType = types.mkOptionType {
+                        name = "haskellOverlay";
+                        description = "An Haskell overlay function";
+                        descriptionClass = "noun";
+                        # NOTE: This check is not exhaustive, as there is no way
+                        # to check that the function takes two arguments, and
+                        # returns an attrset.
+                        check = lib.isFunction;
+                        merge = _loc: defs:
+                          let
+                            logWarning =
+                              if builtins.length defs > 1
+                              then builtins.trace "WARNING[haskell-flake]: Multiple haskell overlays are applied in arbitrary order." null
+                              else null;
+                            overlays =
+                              map (x: x.value)
+                                (builtins.seq
+                                  logWarning
+                                  defs);
+                          in
+                          lib.composeManyExtensions overlays;
+                      };
+                    in
+                    mkOption {
+                      type = haskellOverlayType;
+                      description = ''
+                        Cabal package overrides for this Haskell project
                 
-                      For handy functions, see <https://github.com/NixOS/nixpkgs/blob/master/pkgs/development/haskell-modules/lib/compose.nix>
-                    '';
-                    default = self: super: { };
-                    defaultText = lib.literalExpression "self: super: { }";
-                  };
+                        For handy functions, see 
+                        <https://github.com/NixOS/nixpkgs/blob/master/pkgs/development/haskell-modules/lib/compose.nix>
+
+                        **WARNING**: When using `imports`, multiple overlays
+                        will be merged using `lib.composeManyExtensions`.
+                        However the order the overlays are applied can be
+                        arbitrary (albeit deterministic, based on module system
+                        implementation).  Thus, the use of `overrides` via
+                        `imports` is not officiallly supported. If you'd like
+                        to see proper support, add your thumbs up to
+                        <https://github.com/NixOS/nixpkgs/issues/215486>.
+                      '';
+                      default = self: super: { };
+                      defaultText = lib.literalExpression "self: super: { }";
+                    };
                   packages = mkOption {
                     type = types.lazyAttrsOf packageSubmodule;
                     description = ''