+++ test/flake.nix
@@ -24,19 +24,34 @@
       ];
       perSystem = { self', pkgs, ... }: {
         haskellProjects.default = {
+          # Multiple modules should be merged correctly.
+          imports =
+            let
+              defaults = {
+                overrides = self: super: {
+                  # This is purposefully incorrect (pointing to ./.) because we
+                  # expect it to be overriden below.
+                  foo = self.callCabal2nix "foo" ./. { };
+                };
+                devShell = {
+                  tools = hp: {
+                    # Setting to null should remove this tool from defaults.
+                    ghcid = null;
+                  };
+                  hlsCheck.enable = true;
+                };
+              };
+            in
+            [ defaults ];
           overrides = self: super: {
-            # Custom library overrides (here, "foo" comes from a flake input)
+            # This overrides the overlay above (in `defaults`), because the
+            # module system merges them in such order. cf. the WARNING in option
+            # docs.
             foo = self.callCabal2nix "foo" (inputs.haskell-multi-nix + /foo) { };
           };
-          devShell = {
-            tools = hp: {
-              # Some buildTools are included by default. If you do not want them,
-              # set them to 'null' here.
-              ghcid = null;
-              # You can also add additional build tools.
-              fzf = pkgs.fzf;
-            };
-            hlsCheck.enable = true;
+          devShell.tools = hp: {
+            # Adding a tool should make it available in devshell.
+            fzf = pkgs.fzf;
           };
         };
         # haskell-flake doesn't set the default package, but you can do it here.